modularize profile auth
