+++ README.md
@@ -4,6 +4,9 @@
 
 To run the project locally:
 
+run this the first time: curl -sL https://firebase.tools | bash
+npm install -g firebase-tools
+
 1. Start the Next.js development server:
 
    ```bash
